+++ Makefile
@@ -12,7 +12,6 @@ PIP=$(PYTHON) -m pip
 LINTED=r2ai/code_block.py
 LINTED+=r2ai/bubble.py
 LINTED+=r2ai/const.py
-LINTED+=main.py
 LINTED+=r2ai/backend/kobaldcpp.py
 # LINTED+=r2ai/index.py
 # LINTED+=r2ai/voice.py
