+++ r2ai.r2
@@ -1,3 +1,3 @@
 ?ef [r2ai] Loaded via r2pipe accessible through "$r2ai"
-'$r2ai=#!pipe python main.py
+'$r2ai=#!python r2ai/plugin.py
 # $r2ai -h
