+++ r2ai/cli.py
@@ -1,0 +1,3 @@
+from .main import run # pylint: disable=wrong-import-position
+
+run()
