--- requirements.txt
@@ -1,19 +1,0 @@
-rich
-pyreadline3
-r2pipe
-inquirer
-llama-cpp-python==0.3.1
-huggingface_hub==0.22.2
-appdirs
-unidecode
-jsonref
-transformers
-pydantic
-tokentrim
-chromadb
-openai
-anthropic
-boto3
-colorama
-textual
-litellm
